+++ CHANGELOG.md
@@ -33,6 +33,7 @@ All this means that you need to save the Settings at least once and you should d
 * Added UserRefresh to [main.js](Overlay/main.js)
   * This can be used in custom scripts as a trigger if the most recent are required for the script to function correctly. (See [README.md](README.md))
 * Added skip for repeated alerts
+* Added "below 0" checks for various variables
 * Consolidated the Update functions into UpdateTracker()
 * Consolidated the StartUp functions into StartUp()
 * Consolidated GiftSubs and AnonGiftSubs